[Ref] libmodplug: Silence 64bit warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@8710 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ libopenmpt/libopenmpt_modplug_cpp.cpp
@@ -317,7 +317,7 @@ UINT CSoundFile::GetSampleName( UINT nSample, LPSTR s ) const {
 		}
 	}
 	std::memcpy( s, buf, 32 );
-	return std::strlen( buf );
+	return static_cast<UINT>( std::strlen( buf ) );
 }
 
 UINT CSoundFile::GetInstrumentName( UINT nInstr, LPSTR s ) const {
@@ -334,7 +334,7 @@ UINT CSoundFile::GetInstrumentName( UINT nInstr, LPSTR s ) const {
 		}
 	}
 	std::memcpy( s, buf, 32 );
-	return std::strlen( buf );
+	return static_cast<UINT>( std::strlen( buf ) );
 }
 
 void CSoundFile::LoopPattern( int nPat, int nRow ) {
